feat: Decouples the getGeoLocation function to the utils directory
