+++ modernize/src/app/(DashboardLayout)/apps/orcamento/gerar/page.tsx
@@ -51,7 +51,7 @@ import {
 } from '@mui/material';
 import { calcularDataFuturaDiasUteis, calcDiasNaoUteisEntreDatas } from '@/utils/calcDiasUteis';
 import CustomRadio from '@/app/components/forms/theme-elements/CustomRadio';
-import Address from '@/app/components/frontend-pages/contact/form/Address';
+import getGeoLocation from '@/utils/getGeolocation';
 
 
 interface Cliente {
@@ -529,25 +529,6 @@ const OrcamentoGerarScreen = () => {
       return response.json();
     }
   });
-
-  const getGeoLocation = async (address: string): Promise<{ latitude: number; longitude: number } | null> => {
-    try {
-      const encodedAddress = encodeURIComponent(address);
-      const response = await fetch(`https://nominatim.openstreetmap.org/search?format=json&q=${encodedAddress}`);
-      const data = await response.json();
-
-      if (data && data.length > 0) {
-        const { lat, lon } = data[0];
-        return { latitude: parseFloat(lat), longitude: parseFloat(lon) };
-      } else {
-        console.error('Endereço não encontrado');
-        return null;
-      }
-    } catch (error) {
-      console.error('Erro ao obter localização geográfica:', error);
-      return null;
-    }
-  };
 
   const validateCEP = async (cep: string) => {
